% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{cluster_iteration}
\alias{cluster_iteration}
\title{One cluster iteration}
\usage{
cluster_iteration(clusters)
}
\arguments{
\item{clusters}{data.table of connections with columns `first`, `second` and `cluster`}
}
\value{
A data.table of connections with columns `first`, `second` and `cluster`,  in which the cluster numbers have been updated.
}
\description{
One iteration of finding connected groups of points and labelling them as the same cluster.
Takes a data.table of clusters and merges it with itself to find the next layer of connections,
and relables clusters with the minimum of their two previous cluster numbers.
}
