% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{demo_data}
\alias{demo_data}
\title{Create demo data to test clustering algorithm}
\usage{
demo_data(n = 1000)
}
\arguments{
\item{n}{number of rows in demo data}
}
\value{
A new data.table with `n` rows to feed into the clustering algorithm
}
\description{
Create demo data to test clustering algorithm
}
\examples{
demo_data(n = 1000)
}
