% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points copy.R,
%   R/find_connected_points.R
\name{find_connected_points}
\alias{find_connected_points}
\title{Find connected points less than `eps_t` away from each other in an ordered list.}
\usage{
find_connected_points(dt, eps, eps_t)

find_connected_points(dt, eps, eps_t)
}
\arguments{
\item{dt}{data.table of points with columns `id`, `x`, `y`, `time`.}

\item{eps}{The max distance between two points for them to be "connected".}

\item{eps_t}{The max time between two points for them to be "connected".}

\item{vec}{A sorted vector e.g. of timestamps.}
}
\value{
A data.table with two columns, "first" and "second". These are the indicies of the connected point, and the points it is connected to.

A data.table with two columns, "first" and "second". These are the indicies of
the connected points, and the points it is connected to.
}
\description{
Find connected points less than `eps_t` away from each other in an ordered list.

Find connected points less than `eps` away from each other in space,
and less than `eps_t` away from each other in time, given a data.table of those points
given a list of those points which are sufficiently connected in time.
}
\examples{
find_connected_points(vec = 1:10, eps_t = 1.5)
dt <- data.table(id = 1:100,
                x=rnorm(100),
                y = rnorm(100),
                t = 1:100)
find_connected_points(dt = dt,
               sls = find_temporally_connected_points(vec = dt$t,
                                      eps_t = 1.5),
               eps = 10)
}
