% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{find_equilibrium_clusters}
\alias{find_equilibrium_clusters}
\title{Find final equilibrium clusters}
\usage{
find_equilibrium_clusters(clusters)
}
\arguments{
\item{clusters}{data.table of connections with columns `first`, `second` and `cluster`}
}
\value{
A data.table of connections with columns `first`, `second` and `cluster`,  in which the cluster numbers are final.
}
\description{
Iterates through expanding clusters until finished, then returns the final clusters.
}
