% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{find_spatially_connected_points}
\alias{find_spatially_connected_points}
\title{Find spatially close points}
\usage{
find_spatially_connected_points(dt, sls, eps)
}
\arguments{
\item{dt}{data.table of points with columns `id`, `x`, `y`.}

\item{sls}{The sparse logical similarity matrix output from `find_temporally_connected_points()`.}

\item{eps}{The max distance between two points for them to be "connected".}
}
\value{
A data.table with two columns, "first" and "second". These are the indicies of
the connected points, and the points it is connected to.
}
\description{
Find connected points less than `eps` away from each other in space,
given a list of those points which are sufficiently connected in time.
}
\examples{
dt <- data.table(id = 1:100,
                x=rnorm(100),
                y = rnorm(100),
                t = 1:100)
find_spatially_connected_points(dt = dt,
                      sls = find_temporally_connected_points(vec = dt$t,
                                             eps_t = 1.5),
                      eps = 10)
}
