% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{find_temporally_connected_points}
\alias{find_temporally_connected_points}
\title{Find temporally close points}
\usage{
find_temporally_connected_points(vec, eps_t)
}
\arguments{
\item{vec}{A sorted vector e.g. of timestamps.}

\item{eps_t}{The max distance between two points for them to be "connected".}
}
\value{
A data.table with two columns, "first" and "second". These are the indicies of the connected point, and the points it is connected to.
}
\description{
Find close times less than `eps_t` away from each other in an ordered vector of times.
}
\examples{
find_temporally_connected_points(vec = 1:10, eps_t = 1.5)
}
