% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{get_clusters_from_data}
\alias{get_clusters_from_data}
\title{Get Clusters from Data}
\usage{
get_clusters_from_data(
  df,
  x = "X",
  y = "Y",
  t = "timestamp",
  eps = 1,
  eps_t = 5,
  minpts = 7
)
}
\arguments{
\item{df}{dataframe/data.table/tibble containing trajectories to cluster.}

\item{x}{string, name of the x spatial column}

\item{y}{string, name of the y spatial column}

\item{t}{string, name of the timestamp column}

\item{eps}{Largest distance apart points can be to be directly connected}

\item{eps_t}{Longest time apart points can be to be directly connected}

\item{minpts}{Smallest number of points a point must be connected to, to
not be excluded as a possible source point for connecting further points to a cluster.}
}
\value{
A data.table that looks like `df` but contains two additional columns, `point_density` and `cluster`.
}
\description{
Takes a dataframe/tibble/data.table with identified columns for x, y, t and finds
clusters of points that are connected through being no more than `eps` apart in space,
`eps_t` apart in time, connected through points with a minimum of `min_number` connections.
}
\examples{
dt <- data.table(id = 1:100,
                X = rnorm(100),
                Y = rnorm(100),
                timestamp = 1:100)
get_clusters_from_data <- function(dt
                                  ,x = "X"
                                  ,y = "Y"
                                  ,t = "timestamp"
                                  ,eps = 1 # 1 metre distance threshold from other point
                                  ,eps_t = 5 # 5 second time threshold from other point
                                  ,minpts = 7) # minimum connected to 7 points to continue growing a cluster
}
