% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/find_connected_points.R
\name{remove_noise}
\alias{remove_noise}
\title{Remove noise}
\usage{
remove_noise(reachables, min_number = 4)
}
\arguments{
\item{reachables}{data.table of connections with columns `first`, `second` which
has been output from `find_connected_points()`}

\item{min_number}{The minimum number of connected `second` points for this `first` point to be kept.}
}
\value{
A data.table with two columns, `first` and `second`. These are the indicies of
the connected points, and the points it is connected to.
}
\description{
Remove connections from a 'first' point if that point doesn't have enough points
around it (`min_points`) that it is connected to.
}
\examples{
dt <- data.table(id = 1:100,
                x=rnorm(100),
                y = rnorm(100),
                t = 1:100)
reachables <- find_connected_points(dt = dt,
               sls = find_temporally_connected_points(vec = dt$t,
                                      eps_t = 1.5),
               eps = 10)
remove_noise(reachables = reachables, min_number = 4)
}
